requirements.txt
argparse
kafka_python
flask